Extract a sign-in form component
